+++ CHANGELOG.md
@@ -53,7 +53,8 @@
 - Fix return type of `FileSystemCache::load` to `VmResult<Option<Module>>` in
   order to differentiate missing files from errors.
 - Add in-memory caching for recently used Wasm modules.
-- Rename `CosmCache` to just `cosmwasm_vm::Cache`.
+- Rename `CosmCache` to just `cosmwasm_vm::Cache` and add `CacheOptions` to
+  configure it.
 
 ## 0.11.2 (2020-10-26)